// FE/src/app/shared/models/PlanManger/plan-supplie-detail.model.ts
export class PlanSupplieDetail {
    id?: number;
    quantity?: number;
    symbol?: string;
    techRequired?: string;
    manufacturer?: string;
    deliveryTime?: Date;
    note?: string;
    createdBy?: string;
    updatedBy?: string;
    createdAt?: Date;
    updatedAt?: Date;
    status?: number;

    deviceGroup?: any;
    line?: any;
    supply?: any;
}